+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 # Release notes
 
 # 0.7.4:
+__Deprecated__:
+- **BREAKING CHANGE** Env vars that start with COMET_ are now replaced with SL_ prefix for starlake.cmd
+
 __Feature__:
 - Add SL_PROJECT and SL_TENANT env vars to be used in audit table